refactor : tooltip 수정
