+++ Frontend-main/src/components/common/UserProfile.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { theme } from '../../styles/theme';
+import { theme } from './theme';
 
 const UserProfile = ({ userProfile }) => {
   const profileStyle = {
