+++ Frontend-main/src/components/graph/tooltip/EdgeLabelToggle.jsx
