+++ Frontend-main/src/components/viewer/bookmark/BookmarksPage.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
-import { loadBookmarks, saveBookmarks } from './epub/BookmarkManager';
+import { loadBookmarks, saveBookmarks } from './BookmarkManager';
 
 const bookmarkColors = {
   normal: '#f4f7ff', // 연회색(이전 페이지와 통일)
