--- Frontend-main/src/components/viewer/timeline/timeline.css
@@ -1,309 +1,0 @@
-/* 스포일러 방지 스타일 */
-.spoiler-protected {
-  opacity: 0.3;
-  backdrop-filter: blur(4px);
-  pointer-events: none;
-}
-
-.spoiler-protected::after {
-  content: '🔒';
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  font-size: 1.5rem;
-}
-
-/* 타임라인 특화 애니메이션 - pulse와 fadeOut은 App.css의 enhanced-pulse와 fade-in으로 대체 가능 */
-
-.new-relation {
-  animation: enhanced-pulse 1.5s infinite;
-  border-color: #6C8EFF !important;
-}
-
-.removed-relation {
-  animation: fade-in 0.5s reverse forwards;
-}
-
-/* 챕터 슬라이더 커스텀 스타일 */
-input[type="range"] {
-  -webkit-appearance: none;
-  appearance: none;
-  background: #e2e8f0;
-  height: 4px;
-  border-radius: 2px;
-  outline: none;
-}
-
-input[type="range"]::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  width: 16px;
-  height: 16px;
-  background: #6C8EFF;
-  border-radius: 50%;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-input[type="range"]::-webkit-slider-thumb:hover {
-  transform: scale(1.2);
-  box-shadow: 0 0 0 4px rgba(108, 142, 255, 0.2);
-}
-
-/* 개선된 타임라인 컨테이너 */
-.timeline-container {
-  position: relative;
-  padding-left: 2rem;
-  gap: 2rem;
-}
-
-/* 타임라인 메인 라인 */
-.timeline-container::before {
-  content: '';
-  position: absolute;
-  left: 0.75rem;
-  top: 2rem;
-  bottom: 2rem;
-  width: 3px;
-  background: linear-gradient(to bottom, #6C8EFF 0%, #8B5CF6 50%, #EC4899 100%);
-  border-radius: 2px;
-  box-shadow: 0 0 8px rgba(108, 142, 255, 0.3);
-}
-
-/* 타임라인 아이템 기본 스타일 */
-.timeline-item {
-  position: relative;
-  margin-bottom: 2rem;
-  cursor: pointer;
-  border-radius: 16px;
-  transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-/* 타임라인 노드 (점) */
-.timeline-item::before {
-  content: '';
-  position: absolute;
-  left: -1.25rem;
-  top: 2rem;
-  width: 1.5rem;
-  height: 1.5rem;
-  background: linear-gradient(135deg, #6C8EFF 0%, #8B5CF6 100%);
-  border-radius: 50%;
-  border: 4px solid white;
-  box-shadow: 0 0 0 2px rgba(108, 142, 255, 0.2), 0 4px 12px rgba(108, 142, 255, 0.15);
-  z-index: 2;
-  transition: all 0.3s ease;
-}
-
-/* 잠긴 챕터 노드 */
-.timeline-item.spoiler-protected::before {
-  background: linear-gradient(135deg, #d1d5db 0%, #9ca3af 100%);
-  box-shadow: 0 0 0 2px rgba(209, 213, 219, 0.2);
-}
-
-/* 현재 챕터 노드 */
-.timeline-item.current::before {
-  background: linear-gradient(135deg, #3B82F6 0%, #1D4ED8 100%);
-  transform: scale(1.2);
-  box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.3), 0 6px 20px rgba(59, 130, 246, 0.25);
-  animation: enhanced-pulse 2s ease-in-out infinite;
-}
-
-/* 호버 효과 */
-.timeline-item:hover::before {
-  transform: scale(1.1);
-  box-shadow: 0 0 0 3px rgba(108, 142, 255, 0.4), 0 6px 16px rgba(108, 142, 255, 0.2);
-}
-
-/* 타임라인 시간 뱃지 개선 */
-.timeline-time {
-  display: inline-block;
-  background: linear-gradient(135deg, #f0f7ff 0%, #e6f3ff 100%);
-  padding: 0.5rem 1rem;
-  border-radius: 20px;
-  font-size: 0.9rem;
-  color: #1e40af;
-  font-weight: 600;
-  border: 1px solid rgba(108, 142, 255, 0.2);
-  box-shadow: 0 2px 8px rgba(108, 142, 255, 0.1);
-  transition: all 0.2s ease;
-}
-
-.timeline-time:hover {
-  transform: translateY(-1px);
-  box-shadow: 0 4px 12px rgba(108, 142, 255, 0.15);
-}
-
-/* 태블릿 반응형 (768px ~ 1024px) */
-@media (max-width: 1024px) and (min-width: 768px) {
-  .timeline-container {
-    padding-left: 1.5rem;
-    gap: 1.5rem;
-  }
-  
-  .timeline-item::before {
-    left: -1rem;
-    width: 1.25rem;
-    height: 1.25rem;
-  }
-  
-  .timeline-container::before {
-    left: 0.625rem;
-  }
-}
-
-/* 모바일 반응형 (최대 768px) */
-@media (max-width: 768px) {
-  .timeline-container {
-    padding-left: 1rem;
-    gap: 1rem;
-  }
-  
-  .timeline-container::before {
-    left: 0.375rem;
-    width: 2px;
-    top: 1rem;
-    bottom: 1rem;
-  }
-  
-  .timeline-item {
-    padding: 1rem !important;
-    margin-bottom: 1rem;
-    border-radius: 12px;
-  }
-  
-  .timeline-item::before {
-    left: -0.625rem;
-    top: 1rem;
-    width: 1rem;
-    height: 1rem;
-    border: 2px solid white;
-  }
-  
-  .timeline-item.current::before {
-    transform: scale(1.1);
-  }
-  
-  .timeline-time {
-    font-size: 0.8rem !important;
-    padding: 0.3rem 0.6rem !important;
-    margin-left: 0 !important;
-    margin-top: 0.5rem;
-    display: inline-block;
-  }
-}
-
-/* 작은 모바일 (최대 480px) */
-@media (max-width: 480px) {
-  .timeline-container {
-    padding-left: 0.5rem;
-  }
-  
-  .timeline-container::before {
-    left: 0.25rem;
-  }
-  
-  .timeline-item {
-    padding: 0.75rem !important;
-    font-size: 0.9rem;
-  }
-  
-  .timeline-item::before {
-    left: -0.375rem;
-    width: 0.75rem;
-    height: 0.75rem;
-    top: 0.75rem;
-  }
-  
-  .timeline-time {
-    font-size: 0.75rem !important;
-    padding: 0.2rem 0.5rem !important;
-  }
-  
-  /* 모바일에서 카드 그리드 레이아웃 */
-  .timeline-item {
-    border-left: none !important;
-    border-top: 4px solid #6C8EFF;
-  }
-  
-  .timeline-item.current {
-    border-top-color: #3B82F6;
-  }
-}
-
-.timeline-root {
-  min-height: 100vh;
-  background: #f3f6fa;
-  position: relative;
-}
-
-.timeline-header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 64px;
-  background: #fff;
-  z-index: 100;
-  box-shadow: 0 2px 12px rgba(79,109,222,0.07);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding-left: 2rem;
-  padding-right: 2rem;
-}
-
-.timeline-controls-bar {
-  position: fixed;
-  top: 64px;
-  left: 0;
-  width: 100vw;
-  background: #f8faff;
-  z-index: 90;
-  border-bottom: 1px solid #e5eaf3;
-  display: flex;
-  align-items: center;
-  padding-left: 2rem;
-  padding-right: 2rem;
-  padding-top: 1rem;
-  padding-bottom: 1rem;
-  gap: 2rem;
-}
-
-.timeline-main {
-  margin-top: 144px;
-  /* HEADER_HEIGHT + CONTROLS_HEIGHT */
-  padding-bottom: 2.5rem;
-  overflow-y: auto;
-  min-height: 0;
-}
-
-@media (max-width: 700px) {
-  .timeline-header, .timeline-controls-bar {
-    padding-left: 0.7rem;
-    padding-right: 0.7rem;
-  }
-  .timeline-header {
-    height: 52px;
-    font-size: 1.1rem;
-  }
-  .timeline-controls-bar {
-    top: 52px;
-    padding-top: 0.5rem;
-    padding-bottom: 0.5rem;
-    gap: 0.7rem;
-  }
-  .timeline-main {
-    margin-top: 110px;
-    padding-left: 0.2rem;
-    padding-right: 0.2rem;
-  }
-}
-
-/* 드롭다운 메뉴 스타일 */
-.timeline-controls-bar .relative .absolute {
-  z-index: 200 !important;
-  min-width: 160px;
-  box-shadow: 0 4px 18px rgba(79,109,222,0.13);
-}
