+++ Frontend-main/src/hooks/useTooltipPosition.js
@@ -1,0 +1,125 @@
+import { useState, useEffect, useRef } from 'react';
+
+/**
+ * 툴팁의 드래그 및 위치 계산을 담당하는 커스텀 훅
+ * @param {number} initialX - 초기 X 좌표
+ * @param {number} initialY - 초기 Y 좌표
+ * @returns {object} 위치 상태와 드래그 관련 핸들러들
+ */
+export function useTooltipPosition(initialX, initialY) {
+  const [position, setPosition] = useState({ x: 200, y: 200 });
+  const [showContent, setShowContent] = useState(false);
+  const [isDragging, setIsDragging] = useState(false);
+  const [dragOffset, setDragOffset] = useState({ x: 0, y: 0 });
+  const [hasDragged, setHasDragged] = useState(false);
+  const tooltipRef = useRef(null);
+
+  useEffect(() => {
+    setShowContent(true);
+  }, []);
+
+  const handleMouseDown = (e) => {
+    if (e.target.closest(".tooltip-close-btn")) return;
+    setIsDragging(true);
+    const rect = tooltipRef.current.getBoundingClientRect();
+    setDragOffset({
+      x: e.clientX - rect.left,
+      y: e.clientY - rect.top,
+    });
+  };
+
+  const handleMouseMove = (e) => {
+    if (!isDragging) return;
+    const tooltipRect = tooltipRef.current.getBoundingClientRect();
+    const viewportWidth = Math.min(
+      document.documentElement.clientWidth,
+      window.innerWidth
+    );
+    const viewportHeight = Math.min(
+      document.documentElement.clientHeight,
+      window.innerHeight
+    );
+    const scrollX = window.pageXOffset || document.documentElement.scrollLeft;
+    const scrollY = window.pageYOffset || document.documentElement.scrollTop;
+
+    let newX = e.clientX - dragOffset.x;
+    let newY = e.clientY - dragOffset.y;
+
+    newX = Math.max(
+      scrollX,
+      Math.min(newX, viewportWidth + scrollX - tooltipRect.width)
+    );
+    newY = Math.max(
+      scrollY,
+      Math.min(newY, viewportHeight + scrollY - tooltipRect.height)
+    );
+
+    setPosition({ x: newX, y: newY });
+    setHasDragged(true);
+  };
+
+  const handleMouseUp = () => {
+    setIsDragging(false);
+  };
+
+  // 드래그 이벤트 리스너 등록
+  useEffect(() => {
+    if (isDragging) {
+      window.addEventListener("mousemove", handleMouseMove);
+      window.addEventListener("mouseup", handleMouseUp);
+      document.body.style.userSelect = "none";
+    } else {
+      document.body.style.userSelect = "";
+    }
+    return () => {
+      window.removeEventListener("mousemove", handleMouseMove);
+      window.removeEventListener("mouseup", handleMouseUp);
+      document.body.style.userSelect = "";
+    };
+  }, [isDragging]);
+
+  // 초기 위치 설정
+  useEffect(() => {
+    if (
+      initialX !== undefined &&
+      initialY !== undefined &&
+      tooltipRef.current &&
+      !isDragging &&
+      !hasDragged
+    ) {
+      const tooltipRect = tooltipRef.current.getBoundingClientRect();
+      const viewportWidth = Math.min(
+        document.documentElement.clientWidth,
+        window.innerWidth
+      );
+      const viewportHeight = Math.min(
+        document.documentElement.clientHeight,
+        window.innerHeight
+      );
+      const scrollX = window.pageXOffset || document.documentElement.scrollLeft;
+      const scrollY = window.pageYOffset || document.documentElement.scrollTop;
+
+      let newX = initialX;
+      let newY = initialY;
+
+      newX = Math.max(
+        scrollX,
+        Math.min(newX, viewportWidth + scrollX - tooltipRect.width)
+      );
+      newY = Math.max(
+        scrollY,
+        Math.min(newY, viewportHeight + scrollY - tooltipRect.height)
+      );
+
+      setPosition({ x: newX, y: newY });
+    }
+  }, [initialX, initialY, isDragging, hasDragged]);
+
+  return {
+    position,
+    showContent,
+    isDragging,
+    tooltipRef,
+    handleMouseDown,
+  };
+}
