+++ Frontend-main/src/utils/graphDiff.js
