+++ Frontend-main/src/utils/graphElementUtils.js
